Widen numeric types so big files don't wrap
line_counter in ruby::find_all was u16; release builds wrap on overflow,
so a Ruby file longer than 65535 lines (generated schemas, vendored
db dumps) would silently restart at 0 and credit later mutations to the
wrong line — coverage lookup misses, results garbage. AnalyticsResult's
files_count was i32, forcing main.rs to write files.len().try_into()
.unwrap() with a panic path for the empty case where the cast is
obviously safe. Both are gratuitous narrowing of values that already
fit cleanly in usize/u32.

This change widens line_number to u32 (MutableItem field, find_all
counter, File::extract_line_number return, Coverage::find param) and
files_count to usize (start signature, no more try_into in main).
Existing tests use integer literals that auto-coerce, so no test bodies
needed updates apart from one helper signature in main_test.rs that
took u16 and now takes u32 for consistency.

A property-based test would prove the wrap is gone for any file size,
but the value-of-cost ratio for one is low here. The type system carries
the invariant.

+++ engine/src/analytics/mod.rs
@@ -14,12 +14,12 @@ pub struct MutationResult {
 
 #[derive(Debug, Serialize)]
 pub struct AnalyticsResult {
-    pub files_count: i32,
+    pub files_count: usize,
     pub mutations: Vec<MutationResult>,
 }
 
 impl AnalyticsResult {
-    pub fn start(files_count: i32) -> AnalyticsResult {
+    pub fn start(files_count: usize) -> AnalyticsResult {
         AnalyticsResult {
             files_count,
             mutations: Vec::new(),
